buslens/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
pymysql
pytest
httpx
